abstract Penn DAG to Alma
